+++ backend/canisters/community/api/src/lifecycle/init.rs
@@ -30,7 +30,8 @@ pub struct Args {
     pub default_channel_rules: Option<Rules>,
     pub source_group: Option<SourceGroup>,
     pub video_call_operators: Vec<Principal>,
-    #[serde(with = "serde_bytes")]
+    #[deprecated]
+    #[serde(with = "serde_bytes", default)]
     pub ic_root_key: Vec<u8>,
     pub wasm_version: BuildVersion,
     pub test_mode: bool,