+++ backend/canisters/group/api/src/lifecycle/init.rs
@@ -28,7 +28,8 @@ pub struct Args {
     pub internet_identity_canister_id: CanisterId,
     pub gate_config: Option<AccessGateConfig>,
     pub video_call_operators: Vec<Principal>,
-    #[serde(with = "serde_bytes")]
+    #[deprecated]
+    #[serde(with = "serde_bytes", default)]
     pub ic_root_key: Vec<u8>,
     pub wasm_version: BuildVersion,
     pub test_mode: bool,