+++ backend/canisters/group_index/api/src/lifecycle/init.rs
@@ -13,8 +13,6 @@ pub struct Args {
     pub registry_canister_id: CanisterId,
     pub internet_identity_canister_id: CanisterId,
     pub video_call_operators: Vec<Principal>,
-    #[serde(with = "serde_bytes")]
-    pub ic_root_key: Vec<u8>,
     pub wasm_version: BuildVersion,
     pub test_mode: bool,
 }