+++ backend/canisters/user_index/api/src/lifecycle/init.rs
@@ -21,7 +21,6 @@ pub struct Args {
     pub translations_canister_id: CanisterId,
     pub website_canister_id: CanisterId,
     pub video_call_operators: Vec<Principal>,
-    pub ic_root_key: Vec<u8>,
     pub wasm_version: BuildVersion,
     pub test_mode: bool,
 }