+++ backend/libraries/utils/src/env/canister.rs
@@ -19,6 +19,7 @@ impl CanisterEnv {
 
 thread_local! {
     static CANISTER_ID: LazyCell<CanisterId> = LazyCell::new(ic_cdk::api::canister_self);
+    static ROOT_KEY: LazyCell<Vec<u8>> = LazyCell::new(ic_cdk::api::root_key);
 }
 
 impl Environment for CanisterEnv {
@@ -48,6 +49,10 @@ impl Environment for CanisterEnv {
 
     fn arg_data_raw(&self) -> Vec<u8> {
         ic_cdk::api::msg_arg_data()
+    }
+
+    fn ic_root_key(&self) -> Vec<u8> {
+        ROOT_KEY.with(|rk| rk.to_vec())
     }
 }